Job accept (#33)
* Add job status list frontend

* Frontend for individual job status

* Replace var with let

* Replace let with const

* connecting backend

* Add filesaver dependency

* Increase file size limit

* Add db reset flag

* Connect backend to job info and all jobs

* fix frontend checks

* Add status filter for jobs

* Fix job filter

* Make page status state cleaner

* Cleanup

* Merge with main and cleanup

* Fix Job Upload

* Formatting Fixes

* Remove Duplicate Quesitons File

* Fixes

* Remove address from designerinfo

* cleanup designer info

* Add sampled data generater

* Fix status filter

* Decline redirects to dashboard

* Final cleanup

* Remove duplicate

* Changed to producer query

* Add functionality to accepting

* Add dependency to hook

---------

+++ backend/schema/schema.go
@@ -99,9 +99,9 @@ const (
 
 // Print/printer physical dimensions
 type Dimensions struct {
-	Height uint64 `bson:"height,omitempty" json:"height,omitempty"`
-	Width  uint64 `bson:"width,omitempty" json:"width,omitempty"`
-	Depth  uint64 `bson:"depth,omitempty" json:"depth,omitempty"`
+	Height float64 `bson:"height,omitempty" json:"height,omitempty"`
+	Width  float64 `bson:"width,omitempty" json:"width,omitempty"`
+	Depth  float64 `bson:"depth,omitempty" json:"depth,omitempty"`
 }
 
 // A filament
